Update essential scopes in configuration.md
Adding `users.read`and `users.profile.read` scopes, which appear to be necesseary.  These are listed on the main Readme, but not in the configuration.md file. 

I'm not sure yet if `im.read` is essential or not, although it is also listed in the main readme and not in the configuration.md file.